+++ R/utilities-figures.R
@@ -213,8 +213,8 @@ createEsqlabsPlotGridConfiguration <- function() { # nolint: object_length_linte
   return(plotGridConfiguration)
 }
 
-#' @param projectConfiguration Object of class `ProjectConfiguration`
-#' that contains information about the output paths.
+#' @param outputFolder Path to the folder where the results will be
+#' stored.
 #'
 #' @title Create an instance of `ExportConfiguration` R6 class
 #' @rdname createEsqlabsExportConfiguration
@@ -231,15 +231,15 @@ createEsqlabsPlotGridConfiguration <- function() { # nolint: object_length_linte
 #'
 #' @examples
 #' myProjConfig <- ProjectConfiguration$new()
-#' createEsqlabsExportConfiguration(myProjConfig)
+#' createEsqlabsExportConfiguration(myProjConfig$outputFolder)
 #'
 #' @family create-plotting-configurations
 #'
 #' @export
-createEsqlabsExportConfiguration <- function(projectConfiguration) { # nolint: object_length_linter.
+createEsqlabsExportConfiguration <- function(outputFolder) { # nolint: object_length_linter.
   exportConfiguration <- esqlabsR::ExportConfiguration$new()
 
-  exportConfiguration$path <- projectConfiguration$outputFolder
+  exportConfiguration$path <- outputFolder
   exportConfiguration$dpi <- 300
   # NULL is not supported by ExportConfiguration, so we should assign here
   # something useful. NULL in the ProjectConfiguration currently means "do not
@@ -267,6 +267,10 @@ createEsqlabsExportConfiguration <- function(projectConfiguration) { # nolint: o
 #' specified in the excel sheet will be created. If a plot grid with a given name
 #' does not exist, an error is thrown.
 #'
+#' @param outputFolder Optional - path to the folder where the results will be
+#' stored. If `NULL` (default), `projectConfiguration$outputFolder` is used. Only
+#' relevant for plots specified for export in the `exportConfiguration` sheet.
+#'
 #' @return A list of `ggplot` objects
 #'
 #' @import tidyr
@@ -277,6 +281,7 @@ createPlotsFromExcel <- function(
     simulatedScenarios,
     observedData,
     projectConfiguration,
+    outputFolder = NULL,
     stopIfNotFound = TRUE) {
   validateIsOfType(observedData, "DataSet", nullAllowed = TRUE)
   validateIsOfType(projectConfiguration, "ProjectConfiguration")
@@ -439,7 +444,8 @@ createPlotsFromExcel <- function(
   dfExportConfigurations <- .validateExportConfigurationsFromExcel(dfExportConfigurations, plotGrids)
   if (nrow(dfExportConfigurations) > 0) {
     # create a list of ExportConfiguration objects from dfExportConfigurations
-    defaultExportConfiguration <- createEsqlabsExportConfiguration(projectConfiguration)
+    outputFolder <- outputFolder %||% projectConfiguration$outputFolder
+    defaultExportConfiguration <- createEsqlabsExportConfiguration(outputFolder)
     exportConfigurations <- apply(dfExportConfigurations, 1, \(row){
       exportConfiguration <- .createConfigurationFromRow(
         defaultConfiguration = defaultExportConfiguration,