+++ NEWS.md
@@ -1,5 +1,11 @@
 # esqlabsR (development version)
 
+## Major improvements and new features
+
+- `createPlotsFromExcel` gets a new argument `outputFolder`. If provided, the plots 
+are saved in the specified folder. If `NULL`, default results folder as defined 
+in the `ProjectConfiguration` is used (#495, @PavelBal).
+
 ## Minor improvements and bug fixes
 
 - Function `createScenarios` gets a new argument `stopIfParameterNotFound`.
@@ -9,9 +15,14 @@ simulation. If `FALSE`, non-existing parameters are ignored. (#559, @PavelBal)
 
 - Constructor of a `Scenario` class gets a new argument `stopIfParameterNotFound`.
 
+- Function `saveScenarioResults` returns the path where the results were saved.
+
 ## Breaking changes
 
 - Function `exportSteadyStateToXLS` has been removed in favor of `ospsuite::exportSteadyStateToXLS` (#598, @PavelBal)
+
+- Function `createEsqlabsExportConfiguration` requires a new argument `outputFolder` 
+instead of `projectConfiguration`.
 
 # esqlabsR 5.1.3